refactor: custom 패키지 생성
