+++ app/src/main/java/com/depromeet/sloth/data/network/member/MemberUpdateInfoResponse.kt
@@ -3,9 +3,7 @@ package com.depromeet.sloth.data.network.member
 /**
  * MemberInfoResponse
  *
- * memberId: Int,
  * memberName: "String",
- * email: "String
  */
 
 data class MemberUpdateInfoResponse(
